800_auth/app/requirements.txt
setuptools==59.6.0
flask==2.0.3
requests==2.27.1
python-dotenv==0.20.0
gunicorn==20.1.0